+++ src/collections/types/CollectionUpdate.ts
@@ -48,7 +48,7 @@ export interface CollectionUpdate_collectionUpdate_collection {
   privateMetadata: (CollectionUpdate_collectionUpdate_collection_privateMetadata | null)[];
   backgroundImage: CollectionUpdate_collectionUpdate_collection_backgroundImage | null;
   slug: string;
-  description: any;
+  description: any | null;
   seoDescription: string | null;
   seoTitle: string | null;
 }